fix casting in MinMax value creation

+++ src/main/java/objectCreatorHelper/PayloadFieldCreator.java
@@ -118,10 +118,10 @@ public class PayloadFieldCreator {
 			}
 
 			if (dataType == DataTypeList.int8) {
-				Uint8 i = new Uint8();
-				i.setMin((short) dcpWrapper.getMinForVr(valueReference));
-				i.setMax((short) dcpWrapper.getMaxForVr(valueReference));
-				payloadField.setUint8(i);
+				Int8 i = new Int8();
+				i.setMin((byte) dcpWrapper.getMinForVr(valueReference));
+				i.setMax((byte) dcpWrapper.getMaxForVr(valueReference));
+				payloadField.setInt8(i);
 
 			}
 
@@ -382,10 +382,10 @@ public class PayloadFieldCreator {
 			}
 
 			if (dataType == DataTypeList.int8) {
-				Uint8 i = new Uint8();
-				i.setMin((short) dcpWrapper.getMinForVr(valueReference));
-				i.setMax((short) dcpWrapper.getMaxForVr(valueReference));
-				payloadField.setUint8(i);
+				Int8 i = new Int8();
+				i.setMin((byte) dcpWrapper.getMinForVr(valueReference));
+				i.setMax((byte) dcpWrapper.getMaxForVr(valueReference));
+				payloadField.setInt8(i);
 
 			}
